requirements.txt
aiofiles==23.2.1
aiohappyeyeballs==2.6.1
aiohttp==3.11.14
aiosignal==1.3.2
altair==5.5.0
annotated-types==0.7.0
anyio==4.8.0
async-timeout==4.0.3
attrs==23.2.0
beautifulsoup4==4.13.4
blinker==1.9.0
cachetools==5.5.2
certifi==2024.12.14
cffi==1.16.0
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
cycler==0.10.0
dataclasses-json==0.6.7
distlib==0.3.9
distro==1.9.0
dnspython==2.7.0
elevenlabs==1.50.3
exceptiongroup==1.2.1
faiss-cpu==1.11.0
fastapi==0.115.6
ffmpy==0.5.0
filelock==3.16.1
filetype==1.2.0
Flask==3.1.0
flask-cors==5.0.1
frozenlist==1.5.0
fsspec==2024.12.0
gitdb==4.0.12
GitPython==3.1.44
google-ai-generativelanguage==0.6.15
google-api-core==2.24.2
google-api-python-client==2.166.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.4
googleapis-common-protos==1.69.2
gradio==4.44.1
gradio_client==1.3.0
greenlet==3.1.1
groq==0.15.0
grpcio==1.71.0
grpcio-status==1.71.0
gTTS==2.5.4
h11==0.14.0
httpcore==1.0.7
httplib2==0.22.0
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.27.1
idna==3.10
importlib_metadata==8.6.1
importlib_resources==6.5.2
iniconfig==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.5
jiter==0.9.0
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
kiwisolver==1.3.2
langchain==0.3.25
langchain-community==0.3.20
langchain-core==0.3.58
langchain-google-genai==2.0.10
langchain-groq==0.3.2
langchain-nvidia-ai-endpoints==0.3.9
langchain-text-splitters==0.3.8
langsmith==0.3.18
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.26.1
matplotlib==3.4.3
mdurl==0.1.2
multidict==6.2.0
mypy-extensions==1.0.0
mysql-connector==2.2.9
narwhals==1.39.1
numpy==1.26.4
openai==1.75.0
opencv-python==4.11.0.86
orjson==3.10.14
outcome==1.3.0.post0
packaging==24.2
pandas==2.2.3
pdf2image==1.17.0
pillow==10.4.0
pinecone-client==6.0.0
pinecone-plugin-interface==0.0.7
pipenv==2024.4.1
platformdirs==4.3.7
pluggy==1.5.0
propcache==0.3.0
proto-plus==1.26.1
protobuf==5.29.4
pyarrow==20.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.10.5
pydantic-settings==2.8.1
pydantic_core==2.27.2
pydeck==0.9.1
pydub==0.25.1
Pygments==2.19.1
pymongo==4.12.0
pyparsing==2.4.7
PySocks==1.7.1
pytesseract==0.3.13
pytest==8.3.5
pytest-asyncio==0.26.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.0.1
python-multipart==0.0.20
pytz==2024.2
PyYAML==6.0.2
pypdf==5.5.0
referencing==0.36.2
requests==2.32.3
requests-toolbelt==1.0.0
rich==13.9.4
rpds-py==0.25.0
rsa==4.9
ruff==0.9.1
scikit-learn==1.6.1
scipy==1.13.1
selenium==4.10.0
semantic-version==2.10.0
shapely==2.0.6
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
SpeechRecognition==3.13.0
SQLAlchemy==2.0.39
starlette==0.41.3
streamlit==1.45.1
tenacity==9.0.0
threadpoolctl==3.6.0
toml==0.10.2
tomli==2.2.1
tomlkit==0.12.0
tornado==6.5
tqdm==4.67.1
trio==0.25.1
trio-websocket==0.11.1
typer==0.15.1
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.3.0
uvicorn==0.34.0
virtualenv==20.30.0
watchdog==6.0.0
webdriver-manager==4.0.1
websockets==12.0
Werkzeug==3.1.3
wsproto==1.2.0
yarl==1.18.3
zipp==3.21.0
zstandard==0.23.0
